% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteAsRecode.R
\name{pasteAsFctRecode}
\alias{pasteAsFctRecode}
\title{Paste Clipboard as fct_recode}
\usage{
pasteAsFctRecode()
}
\description{
Call this function as an Addin to insert a recoder based on codebook clipboard
}
