% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteAsRecode.R
\name{use_matcher}
\alias{use_matcher}
\title{Get case_match from vector}
\usage{
use_matcher(vec, numbers = TRUE)
}
\arguments{
\item{vec}{Input vector}

\item{numbers}{Should levels be machted to numbers? (defaults to TRUE)}
}
\description{
Get case_match from vector
}
