% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteAsRecode.R
\name{use_recoder}
\alias{use_recoder}
\title{Get Recoder from Vector}
\usage{
use_recoder(vec, same = TRUE)
}
\arguments{
\item{same}{Should the same factor labels be prepopulated? (defaults to TRUE)}
}
\description{
Get Recoder from Vector
}
